KOGITO-5963: [SW] Jobs Addon based in Messaging / Kafka
