+++ quarkus/addons/jobs/management/deployment/src/main/java/org/kie/kogito/jobs/management/quarkus/deployment/KogitoAddOnJobsManagementProcessor.java
