+++ quarkus/addons/jobs/management/pom.xml
@@ -5,18 +5,17 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.kie.kogito</groupId>
-    <artifactId>kogito-addons-quarkus-parent</artifactId>
+    <artifactId>kogito-addons-quarkus-jobs-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>kogito-addons-quarkus-jobs-parent</artifactId>
-  <name>Kogito Add-Ons Quarkus Jobs Parent</name>
+  <artifactId>kogito-addons-quarkus-jobs-management-parent</artifactId>
+  <name>Kogito Add-Ons Quarkus Jobs Management - Parent</name>
   <packaging>pom</packaging>
 
   <modules>
-    <module>common</module>
-    <module>management</module>
-    <module>messaging</module>
+    <module>runtime</module>
+    <module>deployment</module>
   </modules>
 
 </project>
