+++ quarkus/addons/jobs/management/runtime/src/main/resources/META-INF/quarkus-extension.yaml
