+++ quarkus/addons/jobs/messaging/deployment/src/main/java/org/kie/kogito/jobs/messaging/quarkus/deployment/KogitoAddOnJobsMessagingProcessor.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.kogito.jobs.messaging.quarkus.deployment;
+
+import org.kie.kogito.quarkus.addons.common.deployment.KogitoCapability;
+import org.kie.kogito.quarkus.addons.common.deployment.RequireCapabilityKogitoAddOnProcessor;
+
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+
+class KogitoAddOnJobsMessagingProcessor extends RequireCapabilityKogitoAddOnProcessor {
+
+    private static final String FEATURE = "kogito-addon-jobs-messaging-extension";
+
+    KogitoAddOnJobsMessagingProcessor() {
+        super(KogitoCapability.PROCESSES);
+    }
+
+    @BuildStep
+    FeatureBuildItem feature() {
+        return new FeatureBuildItem(FEATURE);
+    }
+}
