+++ quarkus/addons/jobs/messaging/runtime/src/main/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+jobs-messaging
