+++ quarkus/addons/jobs/messaging/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,12 @@
+name: Kogito Add-On Jobs Service Messaging
+description: Kogito Add-On to interact with the Kogito Jobs Service using events via the reactive messaging api
+metadata:
+  keywords:
+    - kogito
+    - process
+    - bpmn
+    - jobs
+  # TODO: add processes guide
+  guide: https://quarkus.io/guides/kogito
+  categories:
+    - "business-automation"
